:sparkles: feat: add month to year conversion
This commit adds a new logic to the getAge() method that converts the age
value into a decimal if the user inputs the age as a month, such as "6
months" and makes so that the field age can receive decimal numbers
with both a comma and a dot.
It also changes the overall age input constraint logic to the getAge()
method.